% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greyLitScrapR.R
\name{get_articles_from_page}
\alias{get_articles_from_page}
\title{Get all articles from a given page URL}
\usage{
get_articles_from_page(
  url,
  xpath_title = NULL,
  xpath_link = NULL,
  xpath_date = NULL,
  session = NULL
)
}
\arguments{
\item{url}{the URL containing the article lists}

\item{xpath_title}{the xpath to obtain titles}

\item{xpath_link}{the xpath to obtain links}

\item{xpath_date}{the xpath to obtain dates}
}
\value{
a dataframe containing the page URL
The titles, and the links
\preformatted{
data,.frame(page_url, titles, links)
}
}
\description{
Gets the titles, links, and dates of individual articles
from a given page URL
}
