% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{get_elem_value}
\alias{get_elem_value}
\title{gets an element value}
\usage{
get_elem_value(html, xpath)
}
\arguments{
\item{html}{the html to search}

\item{xpath}{the xpath to use}
}
\value{
a list containing all the given values
\pre
}
\description{
Gets the value of an element from a given xpath
}
\keyword{internal}
