% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greyLitScrapR.R, R/private.R
\name{get_page_number}
\alias{get_page_number}
\title{Get page numbers from the URL}
\usage{
get_page_number(url, xpath_page_url, regex_page_number, session = NULL)

get_page_number(url, xpath_page_url, regex_page_number, session = NULL)
}
\arguments{
\item{url}{The URL to scrape}

\item{xpath_page_url}{the xpath to extract the page URL}

\item{regex_page_number}{the regex to extract the page number from
the returned URL}

\item{session}{the session to use, defaulting to no existing session}
}
\value{
a list containing the base URL,
the page number (to append to the base URL),
and the full page URL
\preformatted{list(
   base_url = base_url,
   page_number = page_number,
   page_url = page_url,
   session = scrape$session
 }

a list containing the base URL,
the page number (to append to the base URL),
and the full page URL
\preformatted{list(
   base_url = base_url,
   page_number = page_number,
   page_url = page_url,
   session = scrape$session
 }
}
\description{
Gets page numbers from a site

Gets page numbers from a site
}
\keyword{internal}
