% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{scrape_url}
\alias{scrape_url}
\title{politely scrape a URL}
\usage{
scrape_url(url, session = NULL)
}
\arguments{
\item{url}{The URL to scrape}

\item{session}{the session to use, defaulting to no existing session}
}
\value{
a list containing the scraped URL, and the utilised session
\preformatted{list(html, session)}
}
\description{
Scrapes a URL politely, optionally utilising an existing session
}
\seealso{
\code{\link[polite:bow]{polite::bow()}} to initialise a session,
\code{\link[polite:nod]{polite::nod()}} to change the URL of an existing session,
and \code{\link[polite:scrape]{polite::scrape()}} to scrape the HTML
}
\keyword{internal}
